Feature/review read

+++ src/main/java/com/example/dp/domain/review/service/impl/ReviewServiceImpl.java
@@ -1,28 +1,34 @@
 package com.example.dp.domain.review.service.impl;
 
 import com.example.dp.domain.order.entity.Order;
+import com.example.dp.domain.order.entity.OrderState;
 import com.example.dp.domain.order.repository.OrderRepository;
 import com.example.dp.domain.review.dto.request.ReviewRequestDto;
 import com.example.dp.domain.review.dto.response.ReviewResponseDto;
 import com.example.dp.domain.review.entity.Review;
+import com.example.dp.domain.review.exception.ForbiddenAccessReviewException;
+import com.example.dp.domain.review.exception.ForbiddenCreateReviewException;
+import com.example.dp.domain.review.exception.NotFoundOrderException;
+import com.example.dp.domain.review.exception.NotFoundReviewException;
+import com.example.dp.domain.review.exception.NotFoundUserException;
+import com.example.dp.domain.review.exception.ReviewAlreadyExistsException;
+import com.example.dp.domain.review.exception.ReviewErrorCode;
 import com.example.dp.domain.review.repository.ReviewRepository;
 import com.example.dp.domain.review.service.ReviewService;
-import com.example.dp.domain.review.validator.ReviewValidator;
 import com.example.dp.domain.user.entity.User;
+import com.example.dp.domain.user.repository.UserRepository;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
 @Service
-@Transactional(readOnly = true)
 public class ReviewServiceImpl implements ReviewService {
 
     private final ReviewRepository reviewRepository;
     private final OrderRepository orderRepository;
-
-    private static String NOT_FOUND_ORDER = "해당 주문을 찾을 수 없습니다.";
-    private static String NOT_FOUND_REVIEW = "해당 리뷰를 찾을 수 없습니다.";
+    private final UserRepository userRepository;
 
     @Override
     public ReviewResponseDto createReview(
@@ -30,24 +36,29 @@ public class ReviewServiceImpl implements ReviewService {
         final ReviewRequestDto reviewRequestDto,
         final User user) {
         Order order = orderRepository.findById(orderId)
-            .orElseThrow(() -> new RuntimeException(NOT_FOUND_ORDER));
+            .orElseThrow(() -> new NotFoundOrderException(ReviewErrorCode.NOT_FOUND_ORDER));
 
-        ReviewValidator.validOrderBy(order, user);
-        ReviewValidator.validExistReview(reviewRepository.existsByOrder(order));
+        if(order.getState() != OrderState.COMPLETED){
+            throw new ForbiddenCreateReviewException(ReviewErrorCode.FORBIDDEN_CREATE);
+        }
+
+        if (!order.getUser().getId().equals(user.getId())) {
+            throw new ForbiddenAccessReviewException(ReviewErrorCode.FORBIDDEN_ACCESS);
+        }
+
+        if (reviewRepository.existsByOrder(order)) {
+            throw new ReviewAlreadyExistsException(ReviewErrorCode.EXISTS_REVIEW);
+        }
 
         Review review = Review.builder()
+            .user(user)
             .content(reviewRequestDto.getContent())
             .order(order)
             .build();
 
         review = reviewRepository.save(review);
 
-        return ReviewResponseDto.builder()
-            .id(review.getId())
-            .content(review.getContent())
-            .createdAt(review.getCreatedAt())
-            .modifiedAt(review.getModifiedAt())
-            .build();
+        return toDto(review);
     }
 
     @Override
@@ -57,27 +68,55 @@ public class ReviewServiceImpl implements ReviewService {
         final ReviewRequestDto reviewRequestDto,
         final User user) {
         Review review = reviewRepository.findById(reviewId)
-            .orElseThrow(() -> new RuntimeException(NOT_FOUND_REVIEW));
+            .orElseThrow(() -> new NotFoundReviewException(ReviewErrorCode.NOT_FOUND_REVIEW));
 
-        ReviewValidator.validOrderBy(review.getOrder(), user);
+        if (!review.getOrder().getUser().getId().equals(user.getId())) {
+            throw new ForbiddenAccessReviewException(ReviewErrorCode.FORBIDDEN_ACCESS);
+        }
 
         review.updateContent(reviewRequestDto.getContent());
 
-        return ReviewResponseDto.builder()
-            .id(review.getId())
-            .content(review.getContent())
-            .createdAt(review.getCreatedAt())
-            .modifiedAt(review.getModifiedAt())
-            .build();
+        return toDto(review);
     }
 
     @Override
     public void deleteReview(final Long reviewId, final User user) {
         Review review = reviewRepository.findById(reviewId)
-            .orElseThrow(() -> new RuntimeException(NOT_FOUND_REVIEW));
+            .orElseThrow(() -> new NotFoundReviewException(ReviewErrorCode.NOT_FOUND_REVIEW));
 
-        ReviewValidator.validOrderBy(review.getOrder(), user);
+        if (!review.getOrder().getUser().getId().equals(user.getId())) {
+            throw new ForbiddenAccessReviewException(ReviewErrorCode.FORBIDDEN_ACCESS);
+        }
 
         reviewRepository.delete(review);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<ReviewResponseDto> getUserReviews(final Long userId) {
+        User user = userRepository.findById(userId)
+            .orElseThrow(() -> new NotFoundUserException(ReviewErrorCode.NOT_FOUND_USER));
+
+        return user.getReviews()
+            .stream()
+            .map(this::toDto)
+            .toList();
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<ReviewResponseDto> getAllReviews() {
+        List<Review> reviews = reviewRepository.findAll();
+        return reviews.stream().map(this::toDto).toList();
+    }
+
+    private ReviewResponseDto toDto(Review review){
+        return ReviewResponseDto.builder()
+            .id(review.getId())
+            .writtenBy(review.getUser().getUsername())
+            .content(review.getContent())
+            .createdAt(review.getCreatedAt())
+            .modifiedAt(review.getModifiedAt())
+            .build();
     }
 }
